Rename credential_auth
