+++ src/middleware/authentication.ts
@@ -6,7 +6,7 @@ import * as dotenv from 'dotenv'
 import { withLogto, handleAuthRoutes, LogtoExpressConfig } from '@logto/express'
 import { configLogToExpress } from '../types/constants.js'
 import { AccountAuthHandler } from './auth/account-auth.js'
-import { CredentialAuthHandler } from './auth/credential_auth.js'
+import { CredentialAuthHandler } from './auth/credential-auth.js'
 import { DidAuthHandler } from './auth/did_auth.js'
 import { KeyAuthHandler } from './auth/key_auth.js'
 import { CredentialStatusAuthHandler } from './auth/credential-status_auth.js'
